- refactor: extract tor bridge code (kmp-tor - bisq2 jars) into separate components and use them from the node facade
