+++ androidNode/src/androidMain/kotlin/network/bisq/mobile/android/node/di/AndroidNodeModule.kt
@@ -25,6 +25,7 @@ import network.bisq.mobile.android.node.service.user_profile.NodeUserProfileServ
 import network.bisq.mobile.android.node.service.tor.TorIntegrationService
 import network.bisq.mobile.android.node.service.tor.BisqTorNetworkBridge
 import network.bisq.mobile.android.node.service.tor.TorBootstrapOrchestrator
+import network.bisq.mobile.android.node.service.tor.TorBisqBridge
 import network.bisq.mobile.domain.AndroidUrlLauncher
 import network.bisq.mobile.domain.UrlLauncher
 import network.bisq.mobile.domain.service.accounts.AccountsServiceFacade
@@ -65,7 +66,15 @@ val androidNodeModule = module {
 
     single { AndroidApplicationService.Provider() }
 
-    single<ApplicationBootstrapFacade> { NodeApplicationBootstrapFacade(get(), get(), get(), get()) }
+    single<ApplicationBootstrapFacade> {
+        NodeApplicationBootstrapFacade(
+            get<AndroidApplicationService.Provider>(),
+            get<ConnectivityService>(),
+            get<TorIntegrationService>(),
+            get<TorBootstrapOrchestrator>(),
+            get<TorBisqBridge>()
+        )
+    }
 
     single<MarketPriceServiceFacade> { NodeMarketPriceServiceFacade(get(), get()) }
 
@@ -99,6 +108,9 @@ val androidNodeModule = module {
     }
     single<TorBootstrapOrchestrator> {
         TorBootstrapOrchestrator(get())
+    }
+    single<TorBisqBridge> {
+        TorBisqBridge(get())
     }
 
     single<UrlLauncher> { AndroidUrlLauncher(androidContext()) }
